+++ src/networks/transformer_dlrt.py
@@ -1,7 +1,7 @@
 import numpy as np
 import tensorflow as tf
 
-from networks.dense_layers import DLRTLayerAdaptiveLinear, DLRTLayerAdaptive
+from .dense_layers import DLRTLayerAdaptiveLinear, DLRTLayerAdaptive
 
 # global constants !!!!! DANGEROUS!!!
 MAX_TOKENS = 128
@@ -424,7 +424,7 @@ class Decoder(tf.keras.layers.Layer):
         return low, full
 
 
-class TransformerDLRA(tf.keras.Model):
+class TransformerDLRT(tf.keras.Model):
     def __init__(self, *, num_layers, d_model, num_heads, dff, input_vocab_size,
                  target_vocab_size, rate=0.1, tolerance=0.1):
         super().__init__()
