+++ src/speech_transformer_DLRT.py
@@ -1,4 +1,4 @@
-import networks.transformer_dlrt_fr
+import networks.transformer_dlrt
 import networks.transformer
 
 import tensorflow as tf
@@ -30,9 +30,9 @@ loss_object = tf.keras.losses.SparseCategoricalCrossentropy(
     from_logits=True, reduction='none')
 
 
-def train(low_rank):
-    filename = "./logs/big_DLRA_FR_transformer_f/fix_rank_" + str(low_rank)
-    filename_check = "./weight_checks/big_DLRA_FR_transformer_f/fix_rank_" + str(low_rank)
+def train(tolerance):
+    filename = "./logs/DLRA_transformer_f/tolerance_" + str(tolerance)
+    filename_check = "./weight_checks/DLRA_transformer_f/tolerance_" + str(tolerance)
 
     # load dataset
     examples, metadata = tfds.load('ted_hrlr_translate/pt_to_en', with_info=True, as_supervised=True)
@@ -50,9 +50,9 @@ def train(low_rank):
     train_batches = make_batches(train_examples)
     val_batches = make_batches(val_examples)
 
-    num_layers = 6
-    d_model = 512
-    dff = 2048
+    num_layers = 4
+    d_model = 128
+    dff = 512
     num_heads = 8
     dropout_rate = 0.1
 
@@ -67,7 +67,7 @@ def train(low_rank):
     validation_accuracy = tf.keras.metrics.Mean(name='validation_accuracy')
 
     # build model
-    transformer = networks.transformer_dlra_fr.TransformerDLRAFR(
+    transformer = networks.transformer_dlrt.TransformerDLRT(
         num_layers=num_layers,
         d_model=d_model,
         num_heads=num_heads,
@@ -75,7 +75,7 @@ def train(low_rank):
         input_vocab_size=tokenizers.pt.get_vocab_size().numpy(),
         target_vocab_size=tokenizers.en.get_vocab_size().numpy(),
         rate=dropout_rate,
-        low_rank=low_rank)
+        tolerance=tolerance)
 
     checkpoint_path = filename_check + '/checkpoints'
 
@@ -86,13 +86,12 @@ def train(low_rank):
     ckpt_manager = tf.train.CheckpointManager(ckpt, checkpoint_path, max_to_keep=5)
 
     # if a checkpoint exists, restore the latest checkpoint.
-    if ckpt_manager.latest_checkpoint:
+    if False and ckpt_manager.latest_checkpoint:
         ckpt.restore(ckpt_manager.latest_checkpoint)
         print('Latest checkpoint restored!!')
 
-        # Create logger
+    # Create logger
     log_file, file_name = create_csv_logger_cb(folder_name=filename)
-
     # print headline
     log_string = "epoch;time;loss_train;acc_train;loss_val;acc_val;compression;ranks\n"
     with open(file_name, "a") as log:
@@ -145,8 +144,8 @@ def train(low_rank):
         optimizer.apply_gradients(zip(grads_l_step, transformer.trainable_weights))
 
         # Postprocessing K and L
-        transformer.k_step_postprocessing()
-        transformer.l_step_postprocessing()
+        transformer.k_step_postprocessing_adapt()
+        transformer.l_step_postprocessing_adapt()
 
         # S-Step Preprocessing
         transformer.s_step_preprocessing()
@@ -171,7 +170,6 @@ def train(low_rank):
         tar_real = tar[:, 1:]
 
         predictions, _ = transformer([inp, tar_inp], training=False, step=2)
-
         loss = loss_function(tar_real, predictions)
 
         validation_loss(loss)
@@ -189,6 +187,9 @@ def train(low_rank):
         # inp -> portuguese, tar -> english
         for (batch, (inp, tar)) in enumerate(train_batches):
             train_step_low_rank(inp, tar)
+            # Rank Adaptivity
+            transformer.rank_adaption()
+
             if batch % 50 == 0:
                 print(
                     f'Epoch {epoch + 1} Batch {batch} Loss {train_loss.result():.4f} Accuracy {train_accuracy.result():.4f}')
@@ -273,12 +274,12 @@ if __name__ == '__main__':
     print("Parsing options")
     # --- parse options ---
     parser = OptionParser()
-    parser.add_option("-r", "--low_rank", dest="low_rank", default=50)
+    parser.add_option("-t", "--tolerance", dest="tolerance", default=0.05)
     parser.add_option("-e", "--epochs", dest="epochs", default=500)
 
     (options, args) = parser.parse_args()
-    options.low_rank = int(options.low_rank)
+    options.tolerance = float(options.tolerance)
     options.epochs = int(options.epochs)
     EPOCHS = options.epochs
 
-    train(low_rank=options.low_rank)
+    train(tolerance=options.tolerance)
