+++ src/speech_transformer_DLRT_fr.py
@@ -67,7 +67,7 @@ def train(low_rank):
     validation_accuracy = tf.keras.metrics.Mean(name='validation_accuracy')
 
     # build model
-    transformer = networks.transformer_dlra_fr.TransformerDLRAFR(
+    transformer = networks.transformer_dlrt_fr.TransformerDLRTFR(
         num_layers=num_layers,
         d_model=d_model,
         num_heads=num_heads,
