+++ src/speech_transformer_big_DLRT_fr.py
@@ -1,4 +1,4 @@
-import networks.transformer_dlrt
+import networks.transformer_dlrt_fr
 import networks.transformer
 
 import tensorflow as tf
@@ -30,9 +30,9 @@ loss_object = tf.keras.losses.SparseCategoricalCrossentropy(
     from_logits=True, reduction='none')
 
 
-def train(tolerance):
-    filename = "./logs/DLRA_transformer_f/tolerance_" + str(tolerance)
-    filename_check = "./weight_checks/DLRA_transformer_f/tolerance_" + str(tolerance)
+def train(low_rank):
+    filename = "./logs/big_DLRA_FR_transformer_f/fix_rank_" + str(low_rank)
+    filename_check = "./weight_checks/big_DLRA_FR_transformer_f/fix_rank_" + str(low_rank)
 
     # load dataset
     examples, metadata = tfds.load('ted_hrlr_translate/pt_to_en', with_info=True, as_supervised=True)
@@ -50,9 +50,9 @@ def train(tolerance):
     train_batches = make_batches(train_examples)
     val_batches = make_batches(val_examples)
 
-    num_layers = 4
-    d_model = 128
-    dff = 512
+    num_layers = 6
+    d_model = 512
+    dff = 2048
     num_heads = 8
     dropout_rate = 0.1
 
@@ -67,7 +67,7 @@ def train(tolerance):
     validation_accuracy = tf.keras.metrics.Mean(name='validation_accuracy')
 
     # build model
-    transformer = networks.transformer_dlra.TransformerDLRA(
+    transformer = networks.transformer_dlrt_fr.TransformerDLRTFR(
         num_layers=num_layers,
         d_model=d_model,
         num_heads=num_heads,
@@ -75,7 +75,7 @@ def train(tolerance):
         input_vocab_size=tokenizers.pt.get_vocab_size().numpy(),
         target_vocab_size=tokenizers.en.get_vocab_size().numpy(),
         rate=dropout_rate,
-        tolerance=tolerance)
+        low_rank=low_rank)
 
     checkpoint_path = filename_check + '/checkpoints'
 
@@ -86,12 +86,13 @@ def train(tolerance):
     ckpt_manager = tf.train.CheckpointManager(ckpt, checkpoint_path, max_to_keep=5)
 
     # if a checkpoint exists, restore the latest checkpoint.
-    if False and ckpt_manager.latest_checkpoint:
+    if ckpt_manager.latest_checkpoint:
         ckpt.restore(ckpt_manager.latest_checkpoint)
         print('Latest checkpoint restored!!')
 
-    # Create logger
+        # Create logger
     log_file, file_name = create_csv_logger_cb(folder_name=filename)
+
     # print headline
     log_string = "epoch;time;loss_train;acc_train;loss_val;acc_val;compression;ranks\n"
     with open(file_name, "a") as log:
@@ -144,8 +145,8 @@ def train(tolerance):
         optimizer.apply_gradients(zip(grads_l_step, transformer.trainable_weights))
 
         # Postprocessing K and L
-        transformer.k_step_postprocessing_adapt()
-        transformer.l_step_postprocessing_adapt()
+        transformer.k_step_postprocessing()
+        transformer.l_step_postprocessing()
 
         # S-Step Preprocessing
         transformer.s_step_preprocessing()
@@ -170,6 +171,7 @@ def train(tolerance):
         tar_real = tar[:, 1:]
 
         predictions, _ = transformer([inp, tar_inp], training=False, step=2)
+
         loss = loss_function(tar_real, predictions)
 
         validation_loss(loss)
@@ -187,9 +189,6 @@ def train(tolerance):
         # inp -> portuguese, tar -> english
         for (batch, (inp, tar)) in enumerate(train_batches):
             train_step_low_rank(inp, tar)
-            # Rank Adaptivity
-            transformer.rank_adaption()
-
             if batch % 50 == 0:
                 print(
                     f'Epoch {epoch + 1} Batch {batch} Loss {train_loss.result():.4f} Accuracy {train_accuracy.result():.4f}')
@@ -274,12 +273,12 @@ if __name__ == '__main__':
     print("Parsing options")
     # --- parse options ---
     parser = OptionParser()
-    parser.add_option("-t", "--tolerance", dest="tolerance", default=10)
+    parser.add_option("-r", "--low_rank", dest="low_rank", default=50)
     parser.add_option("-e", "--epochs", dest="epochs", default=500)
 
     (options, args) = parser.parse_args()
-    options.tolerance = float(options.tolerance)
+    options.low_rank = int(options.low_rank)
     options.epochs = int(options.epochs)
     EPOCHS = options.epochs
 
-    train(tolerance=options.tolerance)
+    train(low_rank=options.low_rank)
